+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/singletons/SingletonGenerosJogo.java
@@ -50,8 +50,8 @@ public class SingletonGenerosJogo {
         return generoJogos.contains(updatedGenero);
     }
 
-    public GeneroJogo searchGeneroJogosID(Integer id)
+    public GeneroJogo searchGeneroJogosID(Long id)
     {
-        return generoJogos.get(id);
+        return generoJogos.get(id.intValue());
     }
 }
